% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bios}
\alias{Bios}
\title{Four bioclimatic variables for Denmark.}
\format{
A rasterstack with 4 layers:
\describe{
\item{bio 1}{Annual Mean Temperature}
\item{bio 7}{Temperature Annual Range}
\item{bio 8}{Mean Temperature of Wettest Quarter}
\item{bio 12}{Annual Precipitation}
}
}
\source{
\url{https://www.worldclim.org/}
}
\usage{
Bios
}
\description{
A PackedSpatRaster containing four biolcimatic variables for Denmark
}
\keyword{datasets}
