% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_Points.R
\name{Export_Points}
\alias{Export_Points}
\title{Export created points}
\usage{
Export_Points(Points, ID_Col = "ID", format = "SHP", name = "Samples")
}
\arguments{
\item{Points}{The name of the SF file with the spatial points.}

\item{ID_Col}{Name of the ID column, defaults to "ID".}

\item{format}{a character, it can be SHP, GPX or RDS, default to "SHP".}

\item{name}{The name of the file without the extension, defaults to "Samples".}
}
\value{
An exported SHP, GPX or RDS file
}
\description{
Export created points
}
\examples{
data(Bios)

a <- Stratify(Bios)

library(raster)

plot(a$FinalStack, colNA = "black")

FinalRaster <- a$FinalStack

Points <- Random_Stratified_Min_Dist(ClassRaster = FinalRaster,
                                     MinDist = 2000,
                                     n = 30,
                                     n_to_test = 100)

Export_Points(Points, name = "Selected")
}
