% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Stratified_Min_Dist.R
\name{Random_Stratified_Min_Dist}
\alias{Random_Stratified_Min_Dist}
\title{Stratify a raster Stack into the best fitting classes}
\usage{
Random_Stratified_Min_Dist(
  ClassRaster = NULL,
  MinDist = NULL,
  n = NULL,
  n_to_test = 100
)
}
\arguments{
\item{ClassRaster}{a raster with the classes of your site as integers.}

\item{MinDist}{Minimum distance of a random point to the sampling points and cells with other classes.}

\item{n}{Number of points per class can be one number or a vector specifying de number per class.}

\item{n_to_test}{Number of points to test for the minimum distance.}
}
\value{
An sf with the points and their classes
}
\description{
Stratify a raster Stack into the best fitting classes
}
\examples{
data(Bios)

a <- Stratify(Bios)

library(raster)

plot(a$FinalStack, colNA = "black")

FinalRaster <- a$FinalStack

Points <- Random_Stratified_Min_Dist(ClassRaster = FinalRaster,
                                     MinDist = 2000,
                                     n = 30,
                                     n_to_test = 100)

}
