% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratify.R
\name{Stratify}
\alias{Stratify}
\title{Stratify a raster Stack into the best fitting classes}
\usage{
Stratify(Stack = NULL, LowGroup = 2, HighGroup = 10, Criterion = "calinski")
}
\arguments{
\item{Stack}{a spatraster with all the variables used for generating the classes.}

\item{LowGroup}{an integer number, the lowest number of groups to test for.}

\item{HighGroup}{an integer number, the highest number of groups to test for.}

\item{Criterion}{either calinski or ssi, the criterion used for selecting the best number of groups.}
}
\value{
a list with the raster for the best classes and a dataframe with the best results
}
\description{
Stratify a raster Stack into the best fitting classes
}
\examples{
data(Bios)
Bios <- terra::unwrap(Bios)
a <- Stratify(Bios)

library(terra)

plot(a$FinalStack, colNA = "black")
}
